refactor(message): isolate markdown DOM wiring
