+++ src/components/message-bubble-rewire.test.ts
@@ -9,11 +9,12 @@ import { readFile } from "node:fs/promises";
 // while the chat's repeatedly-rendering MarkdownContent happened to wire them.
 
 const bubble = await readFile(new URL("./message-bubble.tsx", import.meta.url), "utf8");
+const wiring = await readFile(new URL("./message-dom-wiring.ts", import.meta.url), "utf8");
 
-assert.match(bubble, /const observer = new MutationObserver\(\(\) => wireAll\(\)\)/, "a MutationObserver re-runs the wiring");
-assert.match(bubble, /observer\.observe\(el, \{ childList: true, subtree: true \}\)/, "it observes added nodes in the container subtree");
-assert.match(bubble, /return \(\) => observer\.disconnect\(\)/, "the observer is disconnected on cleanup");
-assert.match(bubble, /const wireAll = \(\) => \{[\s\S]*?wireCopyButtons\(el\)[\s\S]*?\};[\s\S]*?wireAll\(\);/, "initial pass wires immediately, then the observer re-wires");
+assert.match(wiring, /const observer = new MutationObserver\(\(\) => wireAll\(\)\)/, "a MutationObserver re-runs the wiring");
+assert.match(wiring, /observer\.observe\(el, \{ childList: true, subtree: true \}\)/, "it observes added nodes in the container subtree");
+assert.match(wiring, /return \(\) => observer\.disconnect\(\)/, "the observer is disconnected on cleanup");
+assert.match(wiring, /const wireAll = \(\) => \{[\s\S]*?wireCopyButtons\(el\)[\s\S]*?\};[\s\S]*?wireAll\(\);/, "initial pass wires immediately, then the observer re-wires");
 
 assert.match(bubble, /ph:thumbs-up/, "assistant action row has thumbs-up");
 assert.match(bubble, /ph:thumbs-down/, "assistant action row has thumbs-down");
